Move files to frontend folder
